+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- Make `fxtest.New` ensure that the app was created successfully. Previously,
+  it would return the app (similar to `fx.New`, which expects the user to verify
+  the error).
 
 ## [1.7.0] - 2018-08-16
 ### Added