+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make `fxtest.New` ensure that the app was created successfully. Previously,
   it would return the app (similar to `fx.New`, which expects the user to verify
   the error).
+- Fix a goroutine leak in `fxtest.Lifecycle`.
 
 ## [1.7.0] - 2018-08-16
 ### Added