+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   it would return the app (similar to `fx.New`, which expects the user to verify
   the error).
 - Update dig container to defer acyclic validation until after Invoke. Application
-  startup time should improve proportional to the side of the dependency graph.
+  startup time should improve proportional to the size of the dependency graph.
 - Fix a goroutine leak in `fxtest.Lifecycle`.
 
 ## [1.7.0] - 2018-08-16